# config/documentation/swagger-admin.json
{
    "swagger": "2.0",
  "info": {
    "title": "Applify Project Development Calculator",
    "description": "Applify Project Development Calculator",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "basePath": "/",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
    "paths": {
...
          "/plateform/addplateform": {
            "post": {
              "summary": "Platefrom Add",
              "tags": ["Plateform"],
              "content": ["application/json"],
              "parameters": [
                {
                  "name": "web",
                  "in": "formData",
                  "description": "Enter how much time for web",
                  "type": "integer"
                },
                {
                  "name": "mobile",
                  "in": "formData",
                  "description": "Enter how much time for mobile",
                  "type": "integer"
                },
                {
                    "name": "both",
                    "in": "formData",
                    "description": "Enter how much time for both",
                    "type": "interger"
                  },
                  {
                    "name": "email",
                    "in": "formData",
                    "description": "Enter how much time for email",     
                    "type": "integer"
                  },
                  {
                    "name": "twoStep",
                    "in": "formData",
                    "description": "Enter how much time for two step verification",
                    "type": "integer"
                  },
                  {
                    "name": "viaSocial",
                    "in": "formData",
                    "description": "Enter how much time for via social login",
                    "type": "integer"
                  },
                  {
                    "name": "NoSignUp",
                    "in": "formData",
                    "description": "Enter how much time for No sign up",
                    "type": "integer"
                  },
                  {
                    "name": "accrequired",
                    "in": "formData",
                    "description": "Enter how much time for user and accout required",
                    "type": "integer"

                  },
                  {
                    "name": "notrequired",
                    "in": "formData",
                    "description": "Enter how much time for not required",
                    "type": "integer"
                  },
                  {
                    "name": "userFeed",
                    "in": "formData",
                    "description": "Enter how much time for user Feed back",
                    "type": "integer"
                  },
                  {
                    "name": "search",
                    "in": "formData",
                    "description": "Enter how much time for search",
                    "type": "integer"               
                  },
                  {
                    "name": "pushNoti",
                    "in": "formData",
                    "description": "Enter how much time for push notification",
                    "type": "integer"
                 
                  },
                  {
                    "name": "messUserMan",
                    "in": "formData",
                    "description": "Enter how much time for message user management",
                    "type": "integer"
            
                  },
                  {
                    "name": "RatingReview",
                    "in": "formData",
                    "description": "Enter how much time for Rating and review",
                    "type": "integer"
                
                  },
                  {
                    "name": "RewardDeal",
                    "in": "formData",
                    "description": "Enter how much time for reward deals",       
                    "type": "integer"
                   
                  },
                  {
                    "name": "payEcommerce",
                    "in": "formData",
                    "description": "Enter how much time for pay and ecommerce",
                    "type": "integer"
                    
                  },
                  {
                    "name": "categoryFilter",
                    "in": "formData",
                    "description": "Enter how much time for category filter",
                    "type": "integer"
              
                  },
                  {
                    "name": "userGenerated",
                    "in": "formData",
                    "description": "Enter how much time for user Generated ",
                    "type": "integer"
       
                  },
                  {
                    "name": "Booking",
                    "in": "formData",
                    "description": "Enter how much time for booking",
                    "type": "integer"
            
                  }
                  
              ],
      
              "responses": {
                "200": {
                  "description": "Category add Successfull",
                  "schema": {
                    "type": "object"
                  }
                },
                "405": {
                  "description": "Invalid input"
                }
              }
            }
          },
          "/plateform/editplateform": {
            "put": {
              "summary": "Platefrom Add",
              "tags": ["Plateform"],
              "content": ["application/json"],
              "parameters": [
                {
                  "name": "id",
                  "in": "formData",
                  "description": "Enter id of plateform",
                  "type": "string",
                  "value":"476e481c-3c75-4798-ac73-3ac6f4f23ac0"
                },
                {
                  "name": "web",
                  "in": "formData",
                  "description": "Enter how much time for web",
                  "type": "integer"
                },
                {
                  "name": "mobile",
                  "in": "formData",
                  "description": "Enter how much time for mobile",
                  "type": "integer"
                },
                {
                    "name": "both",
                    "in": "formData",
                    "description": "Enter how much time for both",
               
                    "type": "interger"
                  },
                  {
                    "name": "email",
                    "in": "formData",
                    "description": "Enter how much time for email",     
                    "type": "integer"
                  },
                  {
                    "name": "twoStep",
                    "in": "formData",
                    "description": "Enter how much time for two step verification",

                    "type": "integer"
                  },
                  {
                    "name": "viaSocial",
                    "in": "formData",
                    "description": "Enter how much time for via social login",
           
                    "type": "integer"
                  },
                  {
                    "name": "NoSignUp",
                    "in": "formData",
                    "description": "Enter how much time for No sign up",
            
                    "type": "integer"
                  },
                  {
                    "name": "accrequired",
                    "in": "formData",
                    "description": "Enter how much time for user and accout required",
               
                    "type": "integer"

                  },
                  {
                    "name": "notrequired",
                    "in": "formData",
                    "description": "Enter how much time for not required",
     
                    "type": "integer"
                  },
                  {
                    "name": "userFeed",
                    "in": "formData",
                    "description": "Enter how much time for user Feed back",
            
                    "type": "integer"
                  },
                  {
                    "name": "search",
                    "in": "formData",
                    "description": "Enter how much time for search",
             
                 "type": "integer"               
                  },
                  {
                    "name": "pushNoti",
                    "in": "formData",
                    "description": "Enter how much time for push notification",
   
                    "type": "integer"
                 
                  },
                  {
                    "name": "messUserMan",
                    "in": "formData",
                    "description": "Enter how much time for message user management",

                    "type": "integer"
            
                  },
                  {
                    "name": "RatingReview",
                    "in": "formData",
                    "description": "Enter how much time for Rating and review",
            
                    "type": "integer"
                
                  },
                  {
                    "name": "RewardDeal",
                    "in": "formData",
                    "description": "Enter how much time for reward deals",
         
                    "type": "integer"
                   
                  },
                  {
                    "name": "payEcommerce",
                    "in": "formData",
                    "description": "Enter how much time for pay and ecommerce",
           
                    "type": "integer"
                    
                  },
                  {
                    "name": "categoryFilter",
                    "in": "formData",
                    "description": "Enter how much time for category filter",
              
                    "type": "integer"
              
                  },
                  {
                    "name": "userGenerated",
                    "in": "formData",
                    "description": "Enter how much time for user Generated ",
                
                    "type": "integer"
       
                  },
                  {
                    "name": "Booking",
                    "in": "formData",
                    "description": "Enter how much time for booking",
     
                    "type": "integer"
            
                  }
                  
              ],
      
              "responses": {
                "200": {
                  "description": "Education add Successfull",
                  "schema": {
                    "type": "object"
                  }
                },
                "405": {
                  "description": "Invalid input"
                }
              }
            }
          },
          "/plateform/getplateform": {
            "get": {
              "x-swagger-router-controller": "middleware-name1",
              "tags": ["Plateform"],
              "description": "Get plateform value List",
              "security": [
                {
                  "Bearer": []
                }
              ],
              "responses": {}
            }
          },
          "/category/addcategory": {
            "post": {
              "summary": "Category Add",
              "tags": ["Category add"],
              "content": ["application/json"],
              "parameters": [
                {
                  "name": "name",
                  "in": "formData",
                  "description": "Enter user Name",
                  "required": false,
                  "type": "string"
                },
                {
                  "name": "email",
                  "in": "formData",
                  "description": "Enter your email",
                  "required": false,
                  "type": "string"
                },
                {
                  "name": "category",
                  "in": "formData",
                  "description": "Enter your category",
                  "required": false,
                  "type": "string"
                }
              ],
      
              "responses": {
                "200": {
                  "description": "Category add Successfull",
                  "schema": {
                    "type": "object"
                  }
                },
                "405": {
                  "description": "Invalid input"
                }
              }
            }
          },
          "/category/editcategory": {
            "put": {
              "summary": "Category Add",
              "tags": ["Category add"],
              "content": ["application/json"],
              "parameters": [
                {
                  "name": "id",
                  "in": "formData",
                  "description": "Enter category id",
                  "required": false,
                  "type": "string"
                },
                {
                  "name": "name",
                  "in": "formData",
                  "description": "Enter user Name",
                  "required": "true",
                  "type": "string"
                },
                {
                  "name": "email",
                  "in": "formData",
                  "description": "Enter your email",
                  "required": false,
                  "type": "string"
                },
                {
                  "name": "category",
                  "in": "formData",
                  "description": "Enter your category",
                  "required": false,
                  "type": "string"
                }
              ],
      
              "responses": {
                "200": {
                  "description": "Education add Successfull",
                  "schema": {
                    "type": "object"
                  }
                },
                "405": {
                  "description": "Invalid input"
                }
              }
            }
          },
          "/category/getcategory": {
            "get": {
              "x-swagger-router-controller": "middleware-name1",
              "tags": ["Category add"],
              "description": "Get all category",
              "security": [
                {
                  "Bearer": []
                }
              ],
              "responses": {}
            }
          },
...
    }
  }